Add /health endpoint
`/health` endpoint basically shows the status of the service. Answers if
there is currently any job that has ever successfully done a refresh. If
yes, then it is healthy, if not, it is not.

This is a bit misleading, because due to the current design, once a job
went to good health, this endpoint will never returns unhealthy again.