Write stdout of check commands to a file
So we can retrieve it for debugging purposes. The current solution might
have a slight chance of race, we are writing to disk, at the same time a
request is made to read the file, we might have a corrupt output. Fixing
this probably introduces too much complexity, and the race is not all that
harmful either. So, I decide against caring about it (at least for now)